+++ src/core/request/utils/const.ts
@@ -1,0 +1,15 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+export const
+	tplRgxp = /\${([^}]+)}/g,
+	dataURIRgxp = /^data:([^;]+);/;
+
+export const
+	isTextType = /^text(?:\/|$)/,
+	isXMLType = /^application\/\w+-xml\b/;